register and unregister user to event

+++ backend/HotvitedbServer/data/SQL scripts/test.sql
@@ -10,3 +10,7 @@ select *
 from event;
 
 delete from event;
+
+delete from event_participant
+where event_id = 'eID'
+AND user_id = 'uID';
